refactor: move bookmarks handlers
