+++ handlers/bookmarks/edit.go
@@ -1,29 +1,30 @@
-package goa4web
+package bookmarks
 
 import (
 	"database/sql"
 	"errors"
-	corecommon "github.com/arran4/goa4web/core/common"
-	common "github.com/arran4/goa4web/handlers/common"
 	"log"
 	"net/http"
 
 	"github.com/arran4/goa4web/core"
+	corecommon "github.com/arran4/goa4web/core/common"
 	"github.com/arran4/goa4web/core/templates"
+	common "github.com/arran4/goa4web/handlers/common"
+	db "github.com/arran4/goa4web/internal/db"
 )
 
-func bookmarksEditPage(w http.ResponseWriter, r *http.Request) {
+func EditPage(w http.ResponseWriter, r *http.Request) {
 	type Data struct {
-		*CoreData
+		*corecommon.CoreData
 		BookmarkContent string
 		Bid             interface{}
 	}
 
 	data := Data{
-		CoreData:        r.Context().Value(common.KeyCoreData).(*CoreData),
+		CoreData:        r.Context().Value(common.KeyCoreData).(*corecommon.CoreData),
 		BookmarkContent: "Category: Example 1\nhttp://www.google.com.au Google\nColumn\nCategory: Example 2\nhttp://www.google.com.au Google\nhttp://www.google.com.au Google\n",
 	}
-	queries := r.Context().Value(common.KeyQueries).(*Queries)
+	queries := r.Context().Value(common.KeyQueries).(*db.Queries)
 	session, ok := core.GetSessionOrFail(w, r)
 	if !ok {
 		return
@@ -52,16 +53,16 @@ func bookmarksEditPage(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-func bookmarksEditSaveActionPage(w http.ResponseWriter, r *http.Request) {
+func EditSaveActionPage(w http.ResponseWriter, r *http.Request) {
 	text := r.PostFormValue("text")
-	queries := r.Context().Value(common.KeyQueries).(*Queries)
+	queries := r.Context().Value(common.KeyQueries).(*db.Queries)
 	session, ok := core.GetSessionOrFail(w, r)
 	if !ok {
 		return
 	}
 	uid, _ := session.Values["UID"].(int32)
 
-	if err := queries.UpdateBookmarks(r.Context(), UpdateBookmarksParams{
+	if err := queries.UpdateBookmarks(r.Context(), db.UpdateBookmarksParams{
 		List: sql.NullString{
 			String: text,
 			Valid:  true,
@@ -76,16 +77,16 @@ func bookmarksEditSaveActionPage(w http.ResponseWriter, r *http.Request) {
 
 }
 
-func bookmarksEditCreateActionPage(w http.ResponseWriter, r *http.Request) {
+func EditCreateActionPage(w http.ResponseWriter, r *http.Request) {
 	text := r.PostFormValue("text")
-	queries := r.Context().Value(common.KeyQueries).(*Queries)
+	queries := r.Context().Value(common.KeyQueries).(*db.Queries)
 	session, ok := core.GetSessionOrFail(w, r)
 	if !ok {
 		return
 	}
 	uid, _ := session.Values["UID"].(int32)
 
-	if err := queries.CreateBookmarks(r.Context(), CreateBookmarksParams{
+	if err := queries.CreateBookmarks(r.Context(), db.CreateBookmarksParams{
 		List: sql.NullString{
 			String: text,
 			Valid:  true,
